refactor: Extract logStreaming code into features
