+++ packages/frontend/editor-ui/src/features/logStreaming.ee/logStreaming.utils.ts
